+++ breeze-core/src/main/kotlin/component/Encrypter.kt
@@ -137,4 +137,37 @@ interface Encrypter {
 
 	//TODO 数字签名
 	//endregion
+
+	companion object{
+		private val encrypters = mutableListOf<Encrypter>()
+
+		/**
+		 * 得到已注册的加密器列表。
+		 */
+		@JvmStatic fun values(): MutableList<Encrypter> {
+			return encrypters
+		}
+
+		/**
+		 * 注册指定的加密器。
+		 */
+		@JvmStatic fun register(encrypter: Encrypter){
+			encrypters.add(encrypter)
+		}
+
+		init {
+			registerDefaultEncrypters()
+		}
+
+		private fun registerDefaultEncrypters(){
+			register(DesEncrypter)
+			register(AesEncrypter)
+			register(Md5Encrypter)
+			register(Sha1Encrypter)
+			register(Sha256Encrypter)
+			register(Sha512Encrypter)
+			register(Sha3256Encrypter)
+			register(Sha3512Encrypter)
+		}
+	}
 }