+++ breeze-core/src/main/kotlin/component/Escaper.kt
@@ -25,95 +25,94 @@ interface Escaper {
 	fun unescape(value: String): String
 
 	//region Default Escapers
+	abstract class AbstractEscaper : Escaper {
+		protected abstract val escapeChars: CharArray
+		protected abstract val escapedStrings: Array<String>
+
+		override fun escape(value: String): String {
+			//这里直接遍历字符数组以提高性能
+			return buildString {
+				val chars = value.toCharArray()
+				for(char in chars) {
+					val index = escapeChars.indexOf(char)
+					if(index == -1) append(char) else append(escapedStrings[index])
+				}
+			}
+		}
+
+		@NotOptimized
+		override fun unescape(value: String): String {
+			var result = value
+			val size = escapeChars.size
+			for(i in 0 until size) {
+				result = result.replace(escapedStrings[i], escapeChars[i].toString())
+			}
+			return result
+		}
+	}
+
 	/**
 	 * Kotlin字符串的转义器。
 	 */
-	object KotlinEscaper : Escaper {
-		private val escapeChars = charArrayOf('\n', '\r', '\t', '\b', '\'', '\"', '\$')
-		private val escapedStrings = arrayOf("\\n", "\\r", "\\t", "\\b", "\\'", "\\\"", "\\\$")
-		override fun escape(value: String) = escapeByPair(value, escapeChars, escapedStrings)
-		override fun unescape(value: String) = unescapeByPair(value, escapeChars, escapedStrings)
+	object KotlinEscaper : AbstractEscaper() {
+		 override val escapeChars = charArrayOf('\n', '\r', '\t', '\b', '\'', '\"', '\$')
+		override val escapedStrings = arrayOf("\\n", "\\r", "\\t", "\\b", "\\'", "\\\"", "\\\$")
 	}
 
 	/**
 	 * Java的转义器。
 	 */
-	object JavaEscaper : Escaper {
-		private val escapeChars = charArrayOf('\n', '\r', '\t', '\b', '\'', '\"')
-		private val escapedStrings = arrayOf("\\n", "\\r", "\\t", "\\b", "\\'", "\\\"")
-		override fun escape(value: String) = escapeByPair(value, escapeChars, escapedStrings)
-		override fun unescape(value: String) = unescapeByPair(value, escapeChars, escapedStrings)
+	object JavaEscaper : AbstractEscaper() {
+		 override val escapeChars = charArrayOf('\n', '\r', '\t', '\b', '\'', '\"')
+		 override val escapedStrings = arrayOf("\\n", "\\r", "\\t", "\\b", "\\'", "\\\"")
 	}
 
 	/**
 	 * JavaScript的转义器。
 	 */
-	object JavaScriptEscaper : Escaper {
-		private val escapeChars = charArrayOf('\n', '\r', '\t', '\b', '\'', '\"')
-		private val escapedStrings = arrayOf("\\n", "\\r", "\\t", "\b", "\\'", "\\\"")
-		override fun escape(value: String) = escapeByPair(value, escapeChars, escapedStrings)
-		override fun unescape(value: String) = unescapeByPair(value, escapeChars, escapedStrings)
+	object JavaScriptEscaper : AbstractEscaper() {
+		override val escapeChars = charArrayOf('\n', '\r', '\t', '\b', '\'', '\"')
+		override val escapedStrings = arrayOf("\\n", "\\r", "\\t", "\b", "\\'", "\\\"")
 	}
 
 	/**
 	 * Json的转义器。
 	 */
-	object JsonEscaper : Escaper {
-		private val escapeChars = charArrayOf('\t', '\b', '\n', '\r', '\"')
-		private val escapedStrings = arrayOf("\\t", "\\b", "\\n", "\\r", "\\\"")
-		override fun escape(value: String) = escapeByPair(value, escapeChars, escapedStrings)
-		override fun unescape(value: String) = unescapeByPair(value, escapeChars, escapedStrings)
+	object JsonEscaper : AbstractEscaper() {
+		override val escapeChars = charArrayOf('\t', '\b', '\n', '\r', '\"')
+		override val escapedStrings = arrayOf("\\t", "\\b", "\\n", "\\r", "\\\"")
 	}
 
 	/**
 	 * Xml的转义器。
 	 */
-	object XmlEscaper : Escaper {
-		private val escapeChars = charArrayOf('<', '>', '&', '\'', '\"')
-		private val escapedStrings = arrayOf("&lt;", "&gt;", "&amp;", "&apos;", "&quot;")
-		override fun escape(value: String) = escapeByPair(value, escapeChars, escapedStrings)
-		override fun unescape(value: String) = unescapeByPair(value, escapeChars, escapedStrings)
+	object XmlEscaper : AbstractEscaper() {
+		override val escapeChars = charArrayOf('<', '>', '&', '\'', '\"')
+		override val escapedStrings = arrayOf("&lt;", "&gt;", "&amp;", "&apos;", "&quot;")
 	}
 
 	/**
 	 * Xml属性的转义器。
 	 */
-	object XmlAttributeEscaper : Escaper {
-		private val escapeChars = charArrayOf('<', '>', '&', '\'', '\"', '\n', '\r', '\t')
-		private val escapedStrings = arrayOf("&lt;", "&gt;", "&amp;", "&apos;", "&quot;", "&#xA;", "&#xD", "&#x9")
-		override fun escape(value: String) = escapeByPair(value, escapeChars, escapedStrings)
-		override fun unescape(value: String) = unescapeByPair(value, escapeChars, escapedStrings)
+	object XmlAttributeEscaper : AbstractEscaper() {
+		override val escapeChars = charArrayOf('<', '>', '&', '\'', '\"', '\n', '\r', '\t')
+		override val escapedStrings = arrayOf("&lt;", "&gt;", "&amp;", "&apos;", "&quot;", "&#xA;", "&#xD", "&#x9")
 	}
 
 	/**
 	 * Xml内容的转义器。
 	 */
-	object XmlContentEscaper : Escaper {
-		private val escapeChars = charArrayOf('<', '>', '&')
-		private val escapedStrings = arrayOf("&lt;", "&gt;", "amp;")
-		override fun escape(value: String) = escapeByPair(value, escapeChars, escapedStrings)
-		override fun unescape(value: String) = unescapeByPair(value, escapeChars, escapedStrings)
+	object XmlContentEscaper : AbstractEscaper() {
+		override val escapeChars = charArrayOf('<', '>', '&')
+		override val escapedStrings = arrayOf("&lt;", "&gt;", "amp;")
 	}
 
 	/**
 	 * Html的转义器。
 	 */
-	object HtmlEscaper : Escaper {
-		private val escapeChars = charArrayOf('<', '>', '&', '\'', '\"')
-		private val escapedStrings = arrayOf("&lt;", "&gt;", "&amp;", "&apos;", "&quot;")
-		override fun escape(value: String) = escapeByPair(value, escapeChars, escapedStrings)
-		override fun unescape(value: String) = unescapeByPair(value, escapeChars, escapedStrings)
-	}
-
-	/**
-	 * 正则表达式的转义器。
-	 *
-	 * 注意：使用`Regex.escape(value)`进行转义，不支持反转义。
-	 */
-	object RegexEscaper : Escaper {
-		override fun escape(value: String) = Regex.escape(value)
-		override fun unescape(value: String) =
-			throw UnsupportedOperationException("Unescape regex is not supported by regex escaper.")
+	object HtmlEscaper : AbstractEscaper() {
+		override val escapeChars = charArrayOf('<', '>', '&', '\'', '\"')
+		override val escapedStrings = arrayOf("&lt;", "&gt;", "&amp;", "&apos;", "&quot;")
 	}
 	//endregion
 
@@ -121,7 +120,7 @@ interface Escaper {
 	/**
 	 * 换行转义器。
 	 *
-	 * 转义换行符`\n`。
+	 * 转义换行符`\n`为`\\n`。
 	 */
 	object LineBreakEscaper : Escaper {
 		override fun escape(value: String) = value.replace("\n", "\\n")
@@ -134,36 +133,47 @@ interface Escaper {
 	 * 转义换行符`\n`为`<br>`。
 	 */
 	object HtmlLineBreakEscaper : Escaper {
-		var escapedTag = "<br>"
+		private const val escapedTag = "<br>"
 		override fun escape(value: String) = value.replace("\n", escapedTag)
 		override fun unescape(value: String) = value.replace(escapedTag, "\n", true)
 	}
 	//endregion
 
 	companion object {
-		//不验证escapeChars和escapedStrings长度是否一致：交由底层负责
+		private val escapers = mutableListOf<Escaper>()
 
-		private fun escapeByPair(value: String, escapeChars: CharArray, escapedStrings: Array<String>): String {
-			//这里直接遍历字符数组以提高性能
-			return buildString {
-				val chars = value.toCharArray()
-				for(char in chars) {
-					val index = escapeChars.indexOf(char)
-					if(index == -1) append(char) else append(escapedStrings[index])
-				}
-			}
+		/**
+		 * 得到已注册的转义器列表。
+		 */
+		@JvmStatic fun values():List<Escaper>{
+			return escapers
 		}
 
-		@NotOptimized
-		private fun unescapeByPair(value: String, escapeChars: CharArray, escapedStrings: Array<String>): String {
-			var result = value
-			val size = escapeChars.size
-			for(i in 0 until size) {
-				result = result.replace(escapedStrings[i], escapeChars[i].toString())
-			}
-			return result
+		/**
+		 * 注册指定的转义器。
+		 */
+		@JvmStatic fun register(escaper: Escaper){
+			escapers.add(escaper)
 		}
 
-		//不需要进行注册
+		init {
+			registerDefaultEscapers()
+			registerLineBreakEscapers()
+		}
+
+		private fun registerDefaultEscapers() {
+			register(KotlinEscaper)
+			register(JavaEscaper)
+			register(JavaScriptEscaper)
+			register(JsonEscaper)
+			register(XmlEscaper)
+			register(XmlAttributeEscaper)
+			register(XmlContentEscaper)
+		}
+
+		private fun registerLineBreakEscapers(){
+			register(LineBreakEscaper)
+			register(HtmlLineBreakEscaper)
+		}
 	}
 }