+++ breeze-core/src/test/kotlin/ExampleTest.kt
@@ -1,6 +1,6 @@
 package com.windea.breezeframework.core
 
-import com.windea.breezeframework.core.component.CaseType.*
+import com.windea.breezeframework.core.component.LetterCase.*
 import com.windea.breezeframework.core.extension.*
 import kotlin.test.*