+++ breeze-dsl/breeze-dsl-cmd-text/src/main/kotlin/CmdTextDsl.kt
@@ -28,8 +28,15 @@ interface CmdTextDsl {
 		abstract val code:String
 		override val inlineText get() = text
 
+		//linux使用echo命令输出时需要加上-e参数，windows不能直接通过echo命令输出
+		//linux: "\e[${code}m${text}\e[0m"
+		//windows: "\u001B[${code}m${text}\u001B[0m"
+		
 		override fun toString(): String {
-			return "\u001B[${code}m${text}\u001B[0m"
+			return when{
+				isWindowsOsType -> "\u001B[${code}m${text}\u001B[0m"
+				else -> "\u001B[${code}m${text}\u001B[0m"
+			}
 		}
 	}
 
@@ -113,5 +120,9 @@ interface CmdTextDsl {
 		Underline(4),
 		Blink(5),
 		Invert(7)
+	}
+	
+	companion object {
+		private val isWindowsOsType = System.getProperty("os.name").contains("windows",true)
 	}
 }