+++ CHANGELOG.md
@@ -447,6 +447,7 @@
 * [X] 整理目录
 * [X] 添加`Iterable.select()`
 * [X] 添加`Validated`，`Ordered`，`Named`
+* [X] 为所有`BreezeComponent`提供静态方法`values`和`register`。
 * [X] breeze-serialization 重构代码
   * [X] 委托实现的serializer
   * [X] serializer不再作为内部类